feat: @AuthenticationPrincipal (#16)
* feat: comment feedback

* feat: AuthenticationPrincipal

* fix: validate functions

* fix: validate nickname

* feat: permit all get methods

+++ src/main/kotlin/com/teamsparta/abrasax/domain/post/comment/service/CommentService.kt
@@ -1,18 +1,23 @@
 package com.teamsparta.abrasax.domain.post.comment.service
 
 
+
+import com.teamsparta.abrasax.common.dto.IdResponseDto
 import com.teamsparta.abrasax.domain.exception.CommentMismatchException
 import com.teamsparta.abrasax.domain.exception.MemberNotFoundException
 import com.teamsparta.abrasax.domain.exception.ModelNotFoundException
+import com.teamsparta.abrasax.domain.exception.UnauthorizedAccessException
 import com.teamsparta.abrasax.domain.member.repository.MemberRepository
 import com.teamsparta.abrasax.domain.post.comment.dto.AddCommentRequestDto
 import com.teamsparta.abrasax.domain.post.comment.dto.CommentResponseDto
 import com.teamsparta.abrasax.domain.post.comment.dto.UpdateCommentRequestDto
 import com.teamsparta.abrasax.domain.post.comment.model.Comment
 import com.teamsparta.abrasax.domain.post.comment.model.toCommentResponseDto
+import com.teamsparta.abrasax.domain.post.comment.model.toIdResponseDto
 import com.teamsparta.abrasax.domain.post.comment.repository.CommentRepository
 import com.teamsparta.abrasax.domain.post.repository.PostRepository
 import org.springframework.data.repository.findByIdOrNull
+import org.springframework.security.core.userdetails.User
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 
@@ -23,40 +28,55 @@ class CommentService(
     private val memberRepository: MemberRepository
 ) {
     @Transactional
-    fun addComment(postId: Long, request: AddCommentRequestDto): CommentResponseDto {
+    fun addComment(user: User, postId: Long, request: AddCommentRequestDto): IdResponseDto {
         val post =
             postRepository.findPostByIdAndDeletedAtIsNull(postId).orElseThrow { ModelNotFoundException("Post", postId) }
-        val (content, authorId) = request
-        val member = memberRepository.findByIdOrNull(authorId) ?: throw MemberNotFoundException(authorId)
+        val (content) = request
+        val userEmail = user.username
+        val member = memberRepository.findByEmail(userEmail) ?: throw MemberNotFoundException(userEmail)
 
         val comment = Comment.of(
             content = content,
             member = member,
             post = post,
         )
-        return commentRepository.save(comment).toCommentResponseDto()
+        return commentRepository.save(comment).toIdResponseDto()
     }
 
     @Transactional
-    fun updateComment(postId: Long, commentId: Long, requestDto: UpdateCommentRequestDto): CommentResponseDto {
 
+    fun updateComment(user: User, postId: Long, commentId: Long, requestDto: UpdateCommentRequestDto): IdResponseDto {
         if (postRepository.existsByIdAndDeletedAtIsNull(postId) == false)
             throw ModelNotFoundException("Post", postId)
         val comment = commentRepository.findByIdAndDeletedAtIsNull(commentId) ?: throw ModelNotFoundException(
             "Comment",
             commentId
         )
-
+        val userEmail = user.username
+        if (comment.member.email != userEmail) throw UnauthorizedAccessException(
+            email = userEmail,
+            modelName = "Comment",
+            id = commentId
+        )
         comment.update(requestDto.content)
-        return comment.toCommentResponseDto()
+        return comment.toIdResponseDto()
     }
 
     @Transactional
-    fun deleteComment(postId: Long, commentId: Long) {
+
+    fun deleteComment(user: User, postId: Long, commentId: Long) {
         val post = postRepository.findByIdOrNull(postId) ?: throw ModelNotFoundException("Post", postId)
         if (post.id != postId) throw CommentMismatchException(postId, commentId)
         val comment = commentRepository.findByIdAndDeletedAtIsNull(commentId)
             ?: throw ModelNotFoundException("Comment", commentId)
+
+        val userEmail = user.username
+        if (comment.member.email != userEmail) throw UnauthorizedAccessException(
+            email = userEmail,
+            modelName = "Comment",
+            id = commentId
+        )
+
         comment.delete()
     }
